+++ src/pages/api/v1/status/index.tsx
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import login from "../../../../../server/controllers/LoginTs";
+import login from "../../../../../server/controllers/Login";
 import ApiError from "../../@types/ApiError";
 
 type Method = "GET";
